backend/requirements.txt
fastapi>=0.118.0
uvicorn[standard]>=0.34.0
pydantic>=2.11.0
openai>=1.54.0
python-dotenv>=1.0.0
pytest>=8.0.0
